moved utils tests to utils folder
