+++ test/utils/withoutTracing.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'bun:test';
-import { isTracing, withoutTracing } from '../src/utils/withoutTracing';
-import { traceFabric } from '../src/traceFabric';
-import { deepClone } from '../src/deepClone';
+import { isTracing, withoutTracing } from '../../src/utils/withoutTracing';
+import { traceFabric } from '../../src/traceFabric';
+import { deepClone } from '../../src/deepClone';
 
 describe('isTracing(...)', () => {
   test('should be true by default', () => {
